pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "py3dmol>=2.5.2",
    "requests>=2.32.5",
    "spyprot>=0.9.6",
    "streamlit>=1.50.0",
]
